+++ src/main/index.ts
@@ -14,6 +14,7 @@ import Store from "electron-store"
 import { is, getResourcePath } from "@main/utils"
 import { startDiscordRPC } from "@main/rpc"
 import { setupGameModeHandlers, startGameDetection, writeGameModeDiagnostic } from "@main/gameMode"
+import { setupSmartOptimizationHandlers } from "@main/smartOptimization"
 
 console.log = log.log
 console.error = log.error
@@ -160,6 +161,7 @@ app
     setupBackupHandlers()
     setupDebloatHandlers()
     setupGameModeHandlers()
+    setupSmartOptimizationHandlers()
     startGameDetection()
     writeGameModeDiagnostic()
     if (store.get("rpcEnabled") !== false) {